requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
elasticsearch==8.15.0